+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 ==============
 
+#### v17.12-rc3
+
+Changes:
+* `./bin/install.txt` compatible `sh`
+* `./lnmp-docker.sh` compatible `sh`
+* Add `secrets` in `docker-compose.yml`
+* Add `secrets` `configs` in `docker-stack.yml`
+Updates:
+* `Redis` 4.0.6
+
 #### v17.12-rc2
 
 Changes: